+++ src/Application/UseCases/Customers/CreateCustomerUseCase.cs
@@ -1,5 +1,5 @@
 ﻿using Application.UseCases.Customers.Validators;
-using Domain.Customer.Model.CustomerAggregate;
+using Domain.Customers.Model.CustomerAggregate;
 using Domain.SeedWork;
 
 namespace Application.UseCases.Customers;
