+++ src/Application/UseCases/Products/GetProductByIdUseCase.cs
@@ -1,7 +1,4 @@
-﻿using Application.UseCases.Customers;
-using Application.UseCases.Products.Validators;
-using Domain.Customer.Model.CustomerAggregate;
-using Domain.Product.ProductAggregate;
+﻿using Domain.Products.ProductAggregate;
 using Domain.SeedWork;
 
 namespace Application.UseCases.Products;
