+++ src/Application/UseCases/Products/Validators/ProductCreationValidator.cs
@@ -1,5 +1,4 @@
-﻿using Domain.Customer.Model.CustomerAggregate;
-using Domain.Product.ProductAggregate;
+﻿using Domain.Products.ProductAggregate;
 
 namespace Application.UseCases.Products.Validators;
