+++ src/Application/UseCases/Products/Validators/ProductPutValidator.cs
@@ -1,4 +1,4 @@
-﻿using Domain.Product.ProductAggregate;
+﻿using Domain.Products.ProductAggregate;
 
 namespace Application.UseCases.Products.Validators;
