--- src/Domain/Order/OrderAggregate/OrderItem.cs
@@ -1,10 +1,0 @@
-﻿using Domain.SeedWork;
-
-
-namespace Domain.Order.OrderAggregate
-{
-    public class OrderItem : Entity
-    {
-        public Products.ProductAggregate.Product Product { get; set; }
-    }
-}
