--- src/Domain/Order/OrderAggregate/OrderValueTotal.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Domain.Order.OrderAggregate
-{
-    public class OrderValueTotal
-    {
-        public double OrderValue { get; set; }
-
-        public OrderValueTotal()
-        {
-            
-        }
-    }
-}
