+++ src/Domain/Orders/OrderAggregate/IOrderRepository.cs
@@ -1,0 +1,9 @@
+﻿using Domain.SeedWork;
+
+namespace Domain.Orders.OrderAggregate
+{
+    public interface IOrderRepository :IRepository<Order>
+    {
+        Task<Order> GetByOrders();
+    }
+}
