+++ src/Domain/Orders/OrderAggregate/Order.cs
@@ -1,19 +1,21 @@
-﻿using Domain.Order.OrderAggregate.Validators;
+﻿using Domain.Customers.Model.CustomerAggregate;
+using Domain.Orders.OrderAggregate.Validators;
+using Domain.Products.ProductAggregate;
 using Domain.SeedWork;
 
-namespace Domain.Order.OrderAggregate
+namespace Domain.Orders.OrderAggregate
 {
     public class Order : Entity, IAggregatedRoot
     {
-        public IList<Products.ProductAggregate.Product> OrderItems {get; set; }
-        public Domain.Customer.Model.CustomerAggregate.Customer? Customer { get; set; }
+        public IList<Product> OrderItems {get; set; }
+        public Customer? Customer { get; set; }
         public OrderStatus Status { get; }
         public DateTime CreateAt { get; }
         public double TotalPrice { get; set; }
 
 
 
-        public Order(Guid id,IList<Products.ProductAggregate.Product> orderItems, Customer.Model.CustomerAggregate.Customer customer)
+        public Order(Guid id,IList<Product> orderItems, Customer customer)
         {
             Id = id;
             OrderItems = orderItems;
