+++ src/Domain/Orders/OrderAggregate/OrderValueTotal.cs
@@ -1,0 +1,7 @@
+﻿namespace Domain.Orders.OrderAggregate
+{
+    public class OrderValueTotal
+    {
+        public double OrderValue { get; set; }
+    }
+}
