+++ src/Domain/Products/ProductAggregate/Category.cs
@@ -1,15 +1,9 @@
-﻿using Domain.Customer.Model.CustomerAggregate.Validators;
-using Domain.Product.ProductAggregate.Validators;
+﻿using Domain.Products.ProductAggregate.Validators;
 using Domain.SeedWork;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Domain.Product.ProductAggregate
+namespace Domain.Products.ProductAggregate
 {
-    public enum ECategory : int
+    public enum ECategory
     {
         Meal = 0,
         Side = 1,
