+++ src/Infrastructure/Persistence/Repositories/BaseRepository.cs
@@ -1,11 +1,5 @@
-﻿using Domain.Product.ProductAggregate;
-using Domain.SeedWork;
+﻿using Domain.SeedWork;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Infrastructure.Persistence.Repositories
 {
