+++ src/Infrastructure/Persistence/Repositories/ProductRepository.cs
@@ -1,6 +1,4 @@
-﻿using Domain.Customer.Model.CustomerAggregate;
-using Domain.Product.ProductAggregate;
-using Domain.Products.ProductAggregate;
+﻿using Domain.Products.ProductAggregate;
 using Microsoft.EntityFrameworkCore;
 
 namespace Infrastructure.Persistence.Repositories;
